+++ prometheus/testutil/testutil.go
@@ -18,13 +18,13 @@ import (
 	"fmt"
 	"io"
 	"reflect"
-	"sort"
 
 	"github.com/prometheus/common/expfmt"
 
 	dto "github.com/prometheus/client_model/go"
 
 	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/internal"
 )
 
 // CollectAndCompare registers the provided Collector with a newly created
@@ -56,7 +56,7 @@ func GatherAndCompare(g prometheus.Gatherer, expected io.Reader, metricNames ...
 		return fmt.Errorf("parsing expected metrics failed: %s", err)
 	}
 
-	if !reflect.DeepEqual(metrics, normalizeMetricFamilies(expectedMetrics)) {
+	if !reflect.DeepEqual(metrics, internal.NormalizeMetricFamilies(expectedMetrics)) {
 		// Encode the gathered output to the readable text format for comparison.
 		var buf1 bytes.Buffer
 		enc := expfmt.NewEncoder(&buf1, expfmt.FmtText)
@@ -69,7 +69,7 @@ func GatherAndCompare(g prometheus.Gatherer, expected io.Reader, metricNames ...
 		// the registry does to spot differences more easily.
 		var buf2 bytes.Buffer
 		enc = expfmt.NewEncoder(&buf2, expfmt.FmtText)
-		for _, mf := range normalizeMetricFamilies(expectedMetrics) {
+		for _, mf := range internal.NormalizeMetricFamilies(expectedMetrics) {
 			if err := enc.Encode(mf); err != nil {
 				return fmt.Errorf("encoding result failed: %s", err)
 			}
@@ -99,65 +99,4 @@ func filterMetrics(metrics []*dto.MetricFamily, names []string) []*dto.MetricFam
 		}
 	}
 	return filtered
-}
-
-// The below sorting code is copied form the Prometheus client library modulo the added
-// label pair sorting.
-// https://github.com/prometheus/client_golang/blob/ea6e1db4cb8127eeb0b6954f7320363e5451820f/prometheus/registry.go#L642-L684
-
-// metricSorter is a sortable slice of *dto.Metric.
-type metricSorter []*dto.Metric
-
-func (s metricSorter) Len() int {
-	return len(s)
-}
-
-func (s metricSorter) Swap(i, j int) {
-	s[i], s[j] = s[j], s[i]
-}
-
-func (s metricSorter) Less(i, j int) bool {
-	sort.Sort(prometheus.LabelPairSorter(s[i].Label))
-	sort.Sort(prometheus.LabelPairSorter(s[j].Label))
-
-	if len(s[i].Label) != len(s[j].Label) {
-		return len(s[i].Label) < len(s[j].Label)
-	}
-
-	for n, lp := range s[i].Label {
-		vi := lp.GetValue()
-		vj := s[j].Label[n].GetValue()
-		if vi != vj {
-			return vi < vj
-		}
-	}
-
-	if s[i].TimestampMs == nil {
-		return false
-	}
-	if s[j].TimestampMs == nil {
-		return true
-	}
-	return s[i].GetTimestampMs() < s[j].GetTimestampMs()
-}
-
-// normalizeMetricFamilies returns a MetricFamily slice with empty
-// MetricFamilies pruned and the remaining MetricFamilies sorted by name within
-// the slice, with the contained Metrics sorted within each MetricFamily.
-func normalizeMetricFamilies(metricFamiliesByName map[string]*dto.MetricFamily) []*dto.MetricFamily {
-	for _, mf := range metricFamiliesByName {
-		sort.Sort(metricSorter(mf.Metric))
-	}
-	names := make([]string, 0, len(metricFamiliesByName))
-	for name, mf := range metricFamiliesByName {
-		if len(mf.Metric) > 0 {
-			names = append(names, name)
-		}
-	}
-	sort.Strings(names)
-	result := make([]*dto.MetricFamily, 0, len(names))
-	for _, name := range names {
-		result = append(result, metricFamiliesByName[name])
-	}
-	return result
 }
